Bug#33797357 Ndb: Reduce suppressed warnings [noclose]
storage/ndb/src/common/portlib/NdbHW.cpp:2837:13: error: variable 'sum_sys' set but not used [-Werror,-Wunused-but-set-variable]
  long long sum_sys = 0;
              ^
              1 error generated.

Also change some long long to uintmax_t and %llu to %ju.

storage/ndb/src/common/mgmcommon/ConfigInfo.cpp:6928:12: error: variable 'data_sz' set but not used [-Werror,-Wunused-but-set-variable]
    Uint64 data_sz = 0;
               ^
               1 error generated.

storage/ndb/src/ndbapi/ClusterMgr.cpp:1082:12: error: variable 'tot_len' set but not used [-Werror,-Wunused-but-set-variable]
    Uint32 tot_len = length;
               ^

storage/ndb/src/ndbapi/trp_buffer.cpp:86:10: error: variable 'sum' set but not used [-Werror,-Wunused-but-set-variable]
  Uint32 sum = 0;
           ^

storage/ndb/src/ndbapi/NdbReceiver.cpp:562:52: error: performing pointer subtraction with a null pointer may have undefined behavior [-Werror,-Wnull-pointer-subtraction]
      (Uint32)(((const Uint32*)pad(pos, 0, bitPos) - (Uint32*)NULL));
                                                         ^ ~~~~~~~~~~~~~

storage/ndb/test/ndbapi/bench/dbPopulate.cpp: In function 'void populateServers(UserHandle*, int)':
storage/ndb/test/ndbapi/bench/dbPopulate.cpp:130:25: error: '%.*d' directive writing between 2 and 10 bytes into a region of size 4 [-Werror=format-overflow=]
  130 |          sprintf(sbuf, "%.*d", SUBSCRIBER_NUMBER_SUFFIX_LENGTH, j);
      |                         ^~~~
storage/ndb/test/ndbapi/bench/dbPopulate.cpp:130:24: note: directive argument in the range [0, 2147483647]
  130 |          sprintf(sbuf, "%.*d", SUBSCRIBER_NUMBER_SUFFIX_LENGTH, j);
      |                        ^~~~~~
storage/ndb/test/ndbapi/bench/dbPopulate.cpp:130:17: note: 'sprintf' output between 3 and 11 bytes into a destination of size 4
  130 |          sprintf(sbuf, "%.*d", SUBSCRIBER_NUMBER_SUFFIX_LENGTH, j);
      |          ~~~~~~~^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

On Windows:

  C:\...\storage\ndb\src\common\util\ndbxfrm_file.cpp(1600,35): warning C4309: 'initializing': truncation of constant value

storage/ndb/src/common/portlib/NdbMem.cpp: In function 'void* NdbMem_AlignedAlloc(size_t, size_t)':
storage/ndb/src/common/portlib/NdbMem.cpp:272:24: error: ignoring return value of 'int posix_memalign(void**, size_t, size_t)', declared with attribute warn_unused_result [-Werror=unused-result]
   (void) posix_memalign(&p, alignment, size);
          ~~~~~~~~~~~~~~^~~~~~~~~~~~~~~~~~~~~

storage/ndb/src/ndbjtie/jtie/test/myapi/myapi.hpp:441:13: warning: non-static data member 'd0' of 'B1' shadows member inherited from type 'B0' [-Wshadow-field]
    int32_t d0;
            ^
storage/ndb/src/ndbjtie/jtie/test/myapi/myapi.hpp:393:13: note: declared here
    int32_t d0;
            ^
storage/ndb/src/ndbjtie/jtie/test/myapi/myapi.hpp:442:19: warning: non-static data member 'd0c' of 'B1'
shadows member inherited from type 'B0' [-Wshadow-field]
    const int32_t d0c;
                  ^
storage/ndb/src/ndbjtie/jtie/test/myapi/myapi.hpp:394:19: note: declared here
    const int32_t d0c;
                  ^

+++ storage/ndb/include/util/ndbxfrm_file.h
@@ -273,9 +273,9 @@ class ndbxfrm_file
              int key_cipher,  // 0 - none, ndb_ndbxfrm1::cipher_*
              int key_selection_mode,   // ndb_ndbxfrm1::key_selection_mode_*
              int key_count,
-             int key_data_unit_size,
-             off_t file_block_size,
-             off_t data_size);
+             size_t key_data_unit_size,
+             size_t file_block_size,
+             Uint64 data_size);
   /*
    * Use abort when you for example has not fulfilled the initialization of the
    * file content and intend to remove the file after close.
@@ -386,14 +386,14 @@ class ndbxfrm_file
   int read_trailer(ndbxfrm_input_reverse_iterator* in);
   int write_header(
       ndbxfrm_output_iterator* out,
-      off_t data_page_size,
+      size_t data_page_size,
       const byte* pwd_key,
       size_t pwd_key_len,
       int kdf_iter_count,
       int key_cipher,  // 0 - none, 1 - cbc, 2 - xts (always no padding)
       int key_selection_mode,   // 0 - same, 1 - pair, 2 - mixed
       int key_count,
-      int key_data_unit_size);
+      size_t key_data_unit_size);
   int write_trailer(ndbxfrm_output_iterator* out,
                     ndbxfrm_output_iterator* extra);
 };
